+++ cronjob/index.js
@@ -1,4 +1,3 @@
-
 import updateAllHotNews from './updateAllHotNews';
 import updatePersonalize from './updatePersonalize';
 import updateHotKeywords from './updateHotKeywords';
@@ -7,7 +6,12 @@ import updateFirstPageForMobile from './updateFirstPageForMobile';
 import updateFirstPageForDesktop from './updateFirstPageForDesktop';
 import updateNewestNews from './updateNewestNews';
 import updateNewestImages from './updateNewestImages';
+import importMNA from './importMNA';
+import importCnaImages from './importCnaImages';
+import importCnaNews from './importCnaNews';
 import importCNYES from './importCNYES';
+
+
 
 module.exports = async () => {
 
@@ -20,6 +24,9 @@ module.exports = async () => {
         updateFirstPageForDesktop.start(),
         updateNewestNews.start(),
         updateNewestImages.start(),
+        importMNA.start(),
+        importCnaImages.start(),
+        importCnaNews.start(),
         importCNYES.start()
     ]);
 };